SRB2: merged r7135:7145 of ALAM
git-svn-id: svn://code.srb2.org/SRB2/trunk@7146 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/screen.h
@@ -81,7 +81,7 @@ typedef struct viddef_s
 	INT32 recalc; // if true, recalc vid-based stuff
 	UINT8 *direct; // linear frame buffer, or vga base mem.
 	INT32 dupx, dupy; // scale 1, 2, 3 value for menus & overlays
-	float fdupx, fdupy; // same as dupx, dupy, but exact value when aspect ratio isn't 320/200
+	INT32/*fixed_t*/ fdupx, fdupy; // same as dupx, dupy, but exact value when aspect ratio isn't 320/200
 	INT32 bpp; // BYTES per pixel: 1 = 256color, 2 = highcolor
 
 	INT32 baseratio; // Used to get the correct value for lighting walls
